+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- use `topicE1` instead of `topic-1` (same with `topic-2` and `topic-3`).
+
 ## [0.3.3] - 2023-07-05
 
 ### Changed